+++ src/stores/index.ts
@@ -3,5 +3,7 @@ import useLoginStore from "./useLoginStore";
 import useModalStore from "./useModalStore";
 import useShowInfoStore from "./useShowInfoStore";
 import useReservationFormStore from "./useReservationFormStore";
+import useImagePreviewUrlStore from "./useImagePreviewUrlStore";
+import useImageFileStore from "./useImageFileStore";
 
-export { useCarouselDragStore, useLoginStore, useModalStore, useReservationFormStore, useShowInfoStore };
+export { useCarouselDragStore, useLoginStore, useModalStore, useReservationFormStore, useShowInfoStore, useImageFileStore, useImagePreviewUrlStore };
